Moved predicates to a separate file
